requirements.txt
flask==2.1.2
markdown==3.3.7
google-generativeai==0.4.1
gunicorn==21.2.0
Werkzeug==2.1.2
